discordbot/requirements.txt
discord.py
aiohttp
websockets
chardet
bs4
requests